requirements.txt
attrs==22.1.0
cycler==0.11.0
future==0.18.2
hdrhistogram==0.9.2
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.4.2
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pbr==5.10.0
Pillow==9.2.0
pluggy==0.13.1
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==6.2.4
python-dateutil==2.8.2
pytz==2022.2.1
scipy==1.9.1
simpy==4.0.1
six==1.16.0
toml==0.10.2
tqdm==4.64.1
Unidecode==1.3.4
